Rename SharedLayout to MainLayout as it is not shared
