+++ src/components/frame/Extensions/layouts/index.ts
@@ -1,3 +1,3 @@
 export { default as ColonyLayout } from './ColonyLayout';
-export { default as SharedLayout } from './SharedLayout';
+export { default as MainLayout } from './MainLayout';
 export { default as UserNavigationWrapper } from './partials/UserNavigationWrapper';
