+++ src/routes/RootRoute.tsx
@@ -7,29 +7,45 @@ import {
   PageThemeContextProvider,
   usePageThemeContext,
 } from '~context';
+import { useAppContext } from '~hooks';
 import { applyTheme } from '~frame/Extensions/themes/utils';
 import { Theme } from '~frame/Extensions/themes/enum';
 import { DialogProvider } from '~shared/Dialog';
+import { isBasicWallet } from '~types';
+import { getLastWallet } from '~utils/autoLogin';
 
 const RootRouteInner = () => {
   const { isDarkMode } = usePageThemeContext();
+  const { wallet, connectWallet } = useAppContext();
+
   useEffect(() => {
     applyTheme(isDarkMode ? Theme.dark : Theme.light);
   }, [isDarkMode]);
+
+  useEffect(() => {
+    if (
+      (!wallet || isBasicWallet(wallet)) &&
+      connectWallet &&
+      getLastWallet()
+    ) {
+      connectWallet();
+    }
+  }, [connectWallet, wallet]);
+
   return (
-    <AppContextProvider>
-      <DialogProvider>
-        <PageHeadingContextProvider>
-          <Outlet />
-        </PageHeadingContextProvider>
-      </DialogProvider>
-    </AppContextProvider>
+    <DialogProvider>
+      <PageHeadingContextProvider>
+        <Outlet />
+      </PageHeadingContextProvider>
+    </DialogProvider>
   );
 };
 
 const RootRoute = () => (
   <PageThemeContextProvider>
-    <RootRouteInner />
+    <AppContextProvider>
+      <RootRouteInner />
+    </AppContextProvider>
   </PageThemeContextProvider>
 );
